added prompt to macos/install.sh

+++ macos/install.sh
@@ -1,5 +1,12 @@
 #!/bin/bash
 
+read -p "This will install a LOT of stuff. Are you sure?" -n 1 -r
+echo "\n" # (optional) move to a new line
+if [[ $REPLY =~ ^[nN]$ ]]
+then
+    exit
+fi
+
 # General
 brew install --cask firefox
 brew install --cask visual-studio-code
